attempting port to 64 bit

+++ hexedit/HexxedCharTableModel.groovy
@@ -28,7 +28,7 @@ class HexxedCharTableModel extends AbstractTableModel {
 		if (max >= hexxedStatus.offset + HexxedConstants.ROWMAX * 16)
 			return HexxedConstants.ROWMAX
 		else
-			return Math.ceil(((max - hexxedStatus.offset)/16)) as Integer
+			return Math.ceil(((max - hexxedStatus.offset)/16)) as Long
 		}
 	}
